pointerenter.html

<!doctype html>
<html>
<!--
	Test cases for Pointer Events v1 spec

	This document references Test Assertions (abbrev TA below) written by Cathy Chan
-->
<head>
	<title>Pointer Events pointerenter Tests</title>
	<meta name="viewport" content="width=device-width">
	<link rel="author" href="mailto:jen.leong@gmail.com">	
	<link rel="help" href="http://www.w3.org/wiki/PointerEvents/TestAssertions">
	<!-- http://w3c-test.org/resources/testharness.js -->

	<script src="./testharness.js"></script>
	<script src="pointer-prefix.js"></script>
	<script>
	var test_pointerenter = async_test("pointerenter event received");

	function run() {
//		setup({ explicit_done: true });				
		var target0 = document.getElementById("target0");
		
//		on_event(target0, pointerPrefix["pointerover"], function(event) {
		on_event(target0, "mouseenter", function(event) {
			test_pointerenter.step(function () {
				assert_true(true, "pointerenter event was received");
			});
			test_pointerenter.done();
	 		
		});
	}

	</script>
	<style>
		div {
			margin: 0em;
			padding: 2em;
		}
		#target0 {
			background: yellow;
			border: 1px solid orange;
		}
		#complete-notice {
			background: #afa;
			border: 1px solid #0a0;
			display: none;
		}
		#pointertype-log {
			font-weight: bold;
		}
	</style>
</head>
<body onload="run()">
	<h1>Pointer Events pointerenter Tests</h1>
	<div id="target0">
		Start with your pointing device outside of this box, then move it over here.
	</div>
	<div id="complete-notice">
		<p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
		<p>Would you like to <a href="#">run the tests again</a> with a different pointer type?</p>
	</div>
	<div id="log"></div>
</body>
</html>
